reporintg without aggregate
