+++ pom.xml
@@ -5,7 +5,7 @@
 
     <groupId>com.github.danielflower.mavenplugins</groupId>
     <artifactId>gitlog-maven-plugin</artifactId>
-    <version>1.11-SNAPSHOT</version>
+    <version>1.12-SNAPSHOT</version>
 
     <packaging>maven-plugin</packaging>
     <name>Maven Git Log Plugin</name>
@@ -69,6 +69,11 @@
             <groupId>org.apache.maven.doxia</groupId>
             <artifactId>doxia-sink-api</artifactId>
             <version>1.0</version>
+        </dependency>
+        <dependency>
+            <groupId>xml-apis</groupId>
+            <artifactId>xml-apis</artifactId>
+            <version>1.3.04</version>
         </dependency>
     </dependencies>
